# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Atlas API",
    "description" : "Operations platform for self-hosted applications.\n\nBase path: `/api/v1`. Authenticate with `Authorization: Bearer <jwt>` from `POST /api/v1/auth/login` or SSO.\n\n**Deprecations**\n- `/api/v1/applications` is a deprecated alias for Project + default Service. Prefer `/api/v1/projects` and `/api/v1/services`. Responses include `Deprecation: true`, `Sunset: Wed, 01 Jul 2027 00:00:00 GMT`, and `Link` successor headers. Alias remains until Sunset; see `docs/api/deprecations.md`.\n\nPublished contract snapshot: `docs/api/openapi.json` (regenerate with `-Datlas.writeOpenApi=true` on `OpenApiContractIntegrationTest`).\n",
    "contact" : {
      "name" : "Atlas",
      "url" : "https://github.com/AdriPardo/atlas"
    },
    "license" : {
      "name" : "Proprietary"
    },
    "version" : "v1"
  },
...
  "servers" : [ {
    "url" : "/",
    "description" : "Current host"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/projects/{projectId}/mail/send" : {
      "post" : {
        "tags" : [ "Project mail" ],
        "summary" : "Send transactional email via platform SMTP relay",
        "operationId" : "send",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-Atlas-Mail-Token",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendProjectMailRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendProjectMailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/mail/provision" : {
      "post" : {
        "tags" : [ "Project mail" ],
        "summary" : "Provision SMTP secrets (smtp.* + mail.api_token) for deploy envFrom",
        "operationId" : "provision",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectMailProvisionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/database/provision" : {
      "post" : {
        "tags" : [ "Project database" ],
        "summary" : "Provision schema + migrator/read roles; store db.url / db.schema secrets",
        "operationId" : "provision_1",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDatabaseProvisionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/mail" : {
      "get" : {
        "tags" : [ "Project mail" ],
        "summary" : "Project mail status (never returns credentials)",
        "operationId" : "status",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectMailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "SendProjectMailRequest" : {
        "type" : "object",
        "properties" : {
          "to" : {
            "type" : "string"
          },
          "cc" : {
            "type" : "string"
          },
          "bcc" : {
            "type" : "string"
          },
          "subject" : {
            "type" : "string"
          },
          "textBody" : {
            "type" : "string"
          },
          "htmlBody" : {
            "type" : "string"
          },
          "apiToken" : {
            "type" : "string"
          }
        },
        "required" : [ "subject", "textBody", "to" ]
      },
      "SendProjectMailResponse" : {
        "type" : "object",
        "properties" : {
          "sent" : {
            "type" : "boolean"
          },
          "detail" : {
            "type" : "string"
          },
          "remainingToday" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ProjectMailProvisionResponse" : {
        "type" : "object",
        "properties" : {
          "from" : {
            "type" : "string"
          },
          "host" : {
            "type" : "string"
          },
          "port" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tls" : {
            "type" : "boolean"
          },
          "rotated" : {
            "type" : "boolean"
          }
        }
      },
...
      "ProjectDatabaseProvisionResponse" : {
        "type" : "object",
        "properties" : {
          "schema" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "databaseName" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "rotated" : {
            "type" : "boolean"
          }
        }
      },
...
      "ProjectMailResponse" : {
        "type" : "object",
        "properties" : {
          "provisionerConfigured" : {
            "type" : "boolean"
          },
          "provisioned" : {
            "type" : "boolean"
          },
          "from" : {
            "type" : "string"
          },
          "host" : {
            "type" : "string"
          },
          "port" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tls" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "remainingSendsToday" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dailySendLimit" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}